- Broke Dallas temp into wrapper - Cleaned up DHT wrapper and renamed - added email to license file
